Fix docker build and move report_builder to cli
