+++ .gitignore
@@ -1,6 +1,4 @@
 .DS_Store
-# common location for local config file containing connection credentials
-server/report_builder/config.yaml
 .vscode/*
 
 # Temp export files
