+++ Dockerfile
@@ -34,8 +34,8 @@ EXPOSE 8000
 FROM base
 
 FROM base as dev
-RUN apt-get update && apt-get install -y curl rsync && \
-    curl -fsSL https://deb.nodesource.com/setup_20.x | bash - && \
+RUN apt-get update && apt-get install -y curl rsync git && \
+    curl -fsSL https://deb.nodesource.com/setup_22.x | bash - && \
     apt-get install -y nodejs && \
     npm install -g yarn && \
     apt-get clean && rm -rf /var/lib/apt/lists/*
