+++ server/.gitignore
@@ -41,9 +41,6 @@ app_data/
 **/*.log
 **/*.gz
 
-# report builder server credentials
-report_builder/config.yaml
-
 # report builder outputs
 /*.html
 /*.xlsx
