+++ server/Cargo.toml
@@ -66,7 +66,6 @@ too_many_arguments = "allow"
 [workspace]
 members = [
   "android",
-  "report_builder",
   "repository",
   "graphql",
   "server",
