+++ server/cli/Cargo.toml
@@ -22,9 +22,6 @@ util = { path = "../util" }
 service = { path = "../service" }
 server = { path = "../server" }
 graphql = { path = "../graphql" }
-report_builder = { path = "../report_builder" }
-
-
 anyhow = { workspace = true }
 async-graphql = { workspace = true }
 thiserror = { workspace = true }
@@ -38,6 +35,7 @@ simple-log = { workspace = true }
 tokio = { workspace = true }
 reqwest = { workspace = true }
 base64 = { workspace = true }
+regex = { workspace = true }
 
 simple_logger = { version = "5.0", features = ["colors"] }
 egui = { version = "0.27" }
