+++ server/cli/src/lib.rs
@@ -1,8 +1,8 @@
 extern crate diesel;
 mod refresh_dates;
 pub use refresh_dates::*;
-mod report_utils;
-pub use report_utils::*;
+mod report;
+pub use report::*;
 mod graphql;
 pub use graphql::*;
