+++ server/cli/src/report/print.rs
@@ -4,28 +4,20 @@ use serde::{Deserialize, Serialize};
 
 use std::{fs, path::Path};
 
-use crate::Format;
+use crate::Api;
 
-const AUTH_QUERY: &str = r#"
-query AuthToken($username: String!, $password: String) {
-  authToken(password: $password, username: $username) {
-    ... on AuthToken {
-      __typename
-      token
-    }
-    ... on AuthTokenError {
-      __typename
-      error {
-        description
-      }
-    }
-  }
+#[derive(clap::ValueEnum, Default, Clone, Serialize)]
+#[serde(rename_all = "UPPERCASE")]
+pub enum Format {
+    #[default]
+    Pdf,
+    Html,
+    Excel,
 }
-"#;
 
 const PRINT_QUERY: &str = r#"
 query GenerateReportDefinition($storeId: String!, $name: String, $report: JSON!, $dataId: String, $arguments: JSON, $format: PrintFormat, $excelTemplate: [Int]) {
-  generateReportDefinition(dataId: $dataId, name: $name, report: $report, storeId: $storeId, arguments: $arguments, format: $format, excelTemplateBuffer: $excelTemplate) {
+  root: generateReportDefinition(dataId: $dataId, name: $name, report: $report, storeId: $storeId, arguments: $arguments, format: $format, excelTemplateBuffer: $excelTemplate) {
     ... on PrintReportNode {
       __typename
       fileId
@@ -47,7 +39,7 @@ query GenerateReportDefinition($storeId: String!, $name: String, $report: JSON!,
 
 const STORES_QUERY: &str = r#"
   query stores($storeName: String) {
-    stores(filter: {name: {equalTo: $storeName}}) {
+    root: stores(filter: {name: {equalTo: $storeName}}) {
       ... on StoreConnector {
         nodes {
           id
@@ -64,71 +56,28 @@ pub struct Config {
     pub password: String,
 }
 
-#[allow(dead_code)]
-#[derive(Debug, Deserialize)]
-struct GraphQlResponse {
-    data: serde_json::Value,
-    errors: Option<serde_json::Value>,
-}
+async fn fetch_store_id(api: &Api, store_name: &str) -> anyhow::Result<String> {
+    let result = api
+        .gql(
+            STORES_QUERY,
+            serde_json::json!({ "storeName": store_name }),
+            None,
+        )
+        .await?;
 
-fn token_request(url: Url, config: &Config) -> anyhow::Result<String> {
-    let body = serde_json::json!({
-      "query": AUTH_QUERY,
-      "variables": {
-        "username": config.username,
-        "password": config.password,
-      }
-    });
-    let result: GraphQlResponse = reqwest::blocking::Client::new()
-        .post(url)
-        .json(&body)
-        .send()?
-        .json()?;
-    let auth_token = &result.data["authToken"];
-    if auth_token["__typename"] != "AuthToken" {
-        return Err(anyhow::Error::msg(format!(
-            "Failed to authenticate: {:?}",
-            result
-        )));
-    }
-    Ok(auth_token["token"].as_str().unwrap().to_string())
-}
-
-fn fetch_store_id(url: Url, token: &str, store_name: &str) -> anyhow::Result<String> {
-    let body = serde_json::json!({
-      "query": STORES_QUERY,
-      "variables": {
-        "storeName": store_name
-      }
-    });
-    let response = reqwest::blocking::Client::new()
-        .post(url)
-        .bearer_auth(token)
-        .json(&body)
-        .send()?;
-    let _status = response.status();
-    let gql_result: GraphQlResponse = response.json()?;
-    let store_id: &Option<&str> = &gql_result
-        .data
-        .get("stores")
-        .and_then(|d| d.get("nodes"))
-        .and_then(|s| s.as_array())
+    let store_id = result
+        .get("nodes")
+        .and_then(|n| n.as_array())
         .and_then(|a| a.first())
         .and_then(|s| s.get("id"))
-        .and_then(|s| s.as_str());
-    let Some(store_id) = store_id else {
-        return Err(anyhow::Error::msg(format!(
-            "Can't find store: {:?}\n{:#?}",
-            store_name, gql_result
-        )));
-    };
+        .and_then(|s| s.as_str())
+        .ok_or_else(|| anyhow::Error::msg(format!("Can't find store: {}", store_name)))?;
 
     Ok(store_id.to_string())
 }
 
-fn generate_request(
-    url: Url,
-    token: &str,
+async fn generate_request(
+    api: &Api,
     store_id: &str,
     name: &Option<String>,
     report: serde_json::Value,
@@ -137,49 +86,40 @@ fn generate_request(
     format: Format,
     excel_template_buffer: Option<Vec<u8>>,
 ) -> anyhow::Result<String> {
-    let body = serde_json::json!({
-      "query": PRINT_QUERY,
-      "variables": {
-        "storeId": store_id,
-        "dataId": data_id,
-        "name": name,
-        "report": report,
-        "arguments": arguments,
-        "format": format,
-        "excelTemplate": excel_template_buffer
-      }
-    });
-    let response = reqwest::blocking::Client::new()
-        .post(url)
-        .bearer_auth(token)
-        .json(&body)
-        .send()?;
-    let status = response.status();
-    let gql_result: GraphQlResponse = response.json()?;
-    let result = &gql_result.data["generateReportDefinition"];
-    if result["__typename"] != "PrintReportNode" {
-        return Err(anyhow::Error::msg(format!(
-            "Failed to generate report: status={:?}  {:#?}",
-            status, gql_result
-        )));
-    }
+    let result = api
+        .gql(
+            PRINT_QUERY,
+            serde_json::json!({
+                "storeId": store_id,
+                "dataId": data_id,
+                "name": name,
+                "report": report,
+                "arguments": arguments,
+                "format": format,
+                "excelTemplate": excel_template_buffer
+            }),
+            Some("PrintReportNode"),
+        )
+        .await?;
 
     let file_id = result["fileId"].as_str().unwrap().to_string();
     Ok(file_id)
 }
 
-fn fetch_file(
-    url: Url,
-    token: &str,
+async fn fetch_file(
+    api: &Api,
     file_id: &str,
     output_filename: &Option<String>,
 ) -> anyhow::Result<String> {
-    let result = reqwest::blocking::Client::new()
+    let url = api.url.join("files")?;
+    let response = reqwest::Client::new()
         .get(url)
-        .bearer_auth(token)
+        .bearer_auth(&api.token)
         .query(&[("id", file_id)])
-        .send()?;
-    let content_disposition = result
+        .send()
+        .await?;
+
+    let content_disposition = response
         .headers()
         .get(reqwest::header::CONTENT_DISPOSITION)
         .ok_or(anyhow::Error::msg("Missing content disposition header"))?
@@ -207,8 +147,8 @@ fn fetch_file(
         }
         None => filename,
     };
-    println!("> Write report to: {}", output_filename);
-    fs::write(&output_filename, result.bytes()?)?;
+    log::info!("Write report to: {}", output_filename);
+    fs::write(&output_filename, response.bytes().await?)?;
     Ok(output_filename)
 }
 
@@ -224,7 +164,7 @@ pub struct ReportGenerateData {
     pub excel_template_buffer: Option<Vec<u8>>,
 }
 
-pub fn generate_report(
+pub async fn generate_report(
     config_path: String,
     store_id: Option<String>,
     store_name: Option<String>,
@@ -236,7 +176,7 @@ pub fn generate_report(
     excel_template_file: Option<String>,
 ) -> anyhow::Result<()> {
     let arguments = if let Some(arguments_file) = arguments_file {
-        println!("> Load arguments from: {}", arguments_file);
+        log::info!("Load arguments from: {}", arguments_file);
         let report_data = fs::read_to_string(arguments_file)
             .map_err(|err| anyhow::Error::msg(format!("Failed to load argument file: {}", err)))?;
         Some(serde_json::from_str(&report_data)?)
@@ -244,7 +184,7 @@ pub fn generate_report(
         None
     };
 
-    println!("> Load report data from: {}", report_file);
+    log::info!("Load report data from: {}", report_file);
     let report_data = fs::read_to_string(report_file).map_err(|err| {
         anyhow::Error::msg(format!("Failed to load report definition file: {}", err))
     })?;
@@ -252,15 +192,15 @@ pub fn generate_report(
         anyhow::Error::msg(format!("Failed to parse report definition file: {}", err))
     })?;
 
-    println!("> Load remote server config from: {}", config_path);
+    log::info!("Load remote server config from: {}", config_path);
     let config_data = fs::read_to_string(config_path)
         .map_err(|err| anyhow::Error::msg(format!("Failed to load config file: {}", err)))?;
-    let config: Config = serde_yaml::from_str(&config_data)
+    let config: Config = serde_yml::from_str(&config_data)
         .map_err(|err| anyhow::Error::msg(format!("Failed to parse config file: {}", err)))?;
 
     let excel_template_buffer = excel_template_file.map(|path| fs::read(path)).transpose()?;
 
-    let inner_data = ReportGenerateData {
+    generate_report_inner(ReportGenerateData {
         report,
         config,
         store_id,
@@ -270,14 +210,11 @@ pub fn generate_report(
         data_id,
         arguments,
         excel_template_buffer,
-    };
-
-    generate_report_inner(inner_data)?;
-
-    Ok(())
+    })
+    .await
 }
 
-pub fn generate_report_inner(input: ReportGenerateData) -> anyhow::Result<()> {
+pub async fn generate_report_inner(input: ReportGenerateData) -> anyhow::Result<()> {
     let ReportGenerateData {
         report,
         config,
@@ -292,17 +229,16 @@ pub fn generate_report_inner(input: ReportGenerateData) -> anyhow::Result<()> {
 
     let base_url = Url::parse(&config.url)
         .map_err(|err| anyhow::Error::msg(format!("Invalid base url: {}", err)))?;
-    let gql_url = base_url.join("graphql")?;
-    let files_url = base_url.join("files")?;
 
-    println!("> User graphql endpoint: {}", gql_url);
-    println!("> Authenticate with remote server");
-    let token = token_request(gql_url.clone(), &config).map_err(|err| {
-        anyhow::Error::msg(format!(
-            "Failed to authenticate with remote server: {}",
-            err
-        ))
-    })?;
+    log::info!("Authenticating with remote server at {}", base_url);
+    let api = Api::new_with_token(base_url, config.username, config.password)
+        .await
+        .map_err(|err| {
+            anyhow::Error::msg(format!(
+                "Failed to authenticate with remote server: {}",
+                err
+            ))
+        })?;
 
     let store_id = if let Some(store_id) = store_id {
         store_id
@@ -312,19 +248,18 @@ pub fn generate_report_inner(input: ReportGenerateData) -> anyhow::Result<()> {
                 "Either store_id or store_name must be specified".to_string(),
             ));
         };
-        println!("> Fetch store id for {store_name}");
-        fetch_store_id(gql_url.clone(), &token, &store_name)?
+        log::info!("Fetching store id for {store_name}");
+        fetch_store_id(&api, &store_name).await?
     };
 
-    println!("> Send report generate request ");
+    log::info!("Sending report generate request");
     let file_name = output_filename.as_ref().and_then(|p| {
         Path::new(&p)
             .file_name()
             .map(|n| n.to_string_lossy().to_string())
     });
     let file_id = generate_request(
-        gql_url.clone(),
-        &token,
+        &api,
         &store_id,
         &file_name,
         report,
@@ -333,10 +268,10 @@ pub fn generate_report_inner(input: ReportGenerateData) -> anyhow::Result<()> {
         format,
         excel_template_buffer,
     )
-    .map_err(|err| anyhow::Error::msg(format!("Failed to fetch report data: {}", err)))?;
+    .await?;
 
-    println!("> Download report from {}", files_url);
-    fetch_file(files_url, &token, &file_id, &output_filename)?;
+    log::info!("Downloading report");
+    fetch_file(&api, &file_id, &output_filename).await?;
 
     Ok(())
 }
