--- server/report_builder/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "report_builder"
-version = "0.1.0"
-edition = "2018"
-
-[lib]
-path = "src/lib.rs"
-doctest = false
-
-[[bin]]
-path = "src/cli.rs"
-name = "report_builder"
-test = false
-bench = false
-
-[dependencies]
-service = { path = "../service" }
-
-anyhow.workspace = true
-clap = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-serde_yaml = { workspace = true }
-log = { workspace = true }
-reqwest = { workspace = true }
-regex = { workspace = true }
-base64 = "0.22.1"
