--- server/report_builder/example/file.html
@@ -1,1 +1,0 @@
-Content from file.html
