--- server/report_builder/example/footer.html
@@ -1,3 +1,0 @@
-<div style="font-size: 10px; padding-top: 8px; text-align: center; width: 100%;">
-</span>Some footer here... Page number: <span class="pageNumber"></span>
-</div>
