--- server/report_builder/example/header.html
@@ -1,3 +1,0 @@
-<div style="font-size: 10px; padding-top: 8px; text-align: center; width: 100%;">
-  <span>Some header here...</span>
-</div>
