--- server/report_builder/example/query.graphql
@@ -1,15 +1,0 @@
-query StocktakeQuery($storeId: String!, $dataId: String!) {
-  stocktake(storeId: $storeId, id: $dataId) {
-    ... on StocktakeNode {
-      id
-      comment
-      createdDatetime
-    }
-    ... on NodeError {
-      __typename
-      error {
-        description
-      }
-    }
-  }
-}
