--- server/report_builder/example/report-manifest.json
@@ -1,3 +1,0 @@
-{
-  "context": "STOCKTAKE"
-}
