--- server/report_builder/example/resource_data.json
@@ -1,4 +1,0 @@
-{
-  "value": 5,
-  "value2": "data"
-}
