--- server/report_builder/example/template.html
@@ -1,58 +1,0 @@
-{% set_global timezone = "Asia/Kathmandu" %}
-
-<style>
-  {% include "style.css" %}
-</style>
-
-<body>
-
-<h1>Example Stocktake Report</h1>
-
-The following sections show some example how to access data and resources, include other template files...
-
-<h2>Display a base64 encoded logo</h2>
-The logo data is stored as a data uri scheme in the logo_msupply file so that it can be used directly in an img tag:
-<img style='display:block; width: 100px;' src="{% include "logo_msupply" %}"/>
-
-<h2>Access query data</h2>
-In the Tera template you have access to a data object which contain the results from the data query.
-<br/>
-For example, the data object for the query (for this stocktake report), looks like:
-<pre>
-"data": {{ data | json_encode(pretty=true) }}
-</pre>
-
-<h3>Use data in report</h3>
-Stocktake:
-<table>
-  <tr>
-    <th>id</th>
-    <th>createdDatetime (in local timezone)</th>
-    <th>comment</th>
-  </tr>
-  <tr>
-    <td>{{data.stocktake.id}}</td>
-    <td>
-      {{data.stocktake.createdDatetime | date(format="%Y-%m-%d %H:%M",timezone=timezone)}}
-    </td>
-    <td>{{data.stocktake.comment}}</td>
-  </tr>
-</table>
-
-Note, the query result form the remote-server contains UTC datetimes.
-The example above converts this UTC datetime to a local datetime.
-The used timezone is defined as a Tera variable at the very top of this file.
-
-<h2>Access resource data</h2>
-The report definition contains Tera templates and an arbitrary list of other resources.
-These resources can be accessed through the res object.
-
-The res object in this report is:
-<pre>
-"res": {{ res | json_encode(pretty=true) }}
-</pre>
-
-<h2>Include a different file:</h2>
-{% include "file.html" %}
-
-</body>
