--- server/report_builder/example_arguments/query.graphql
@@ -1,15 +1,0 @@
-query MyQuery($storeId: String!, $creationDatetimeBeforeOrEqual: String!) {
-  invoices(
-    storeId: $storeId
-    filter: {
-      createdDatetime: { beforeOrEqualTo: $creationDatetimeBeforeOrEqual }
-    }
-  ) {
-    ... on InvoiceConnector {
-      nodes {
-        comment
-        id
-      }
-    }
-  }
-}
